+++ src/app/instituciones/colegios/Primaria/page.tsx
@@ -1,0 +1,2 @@
+
+
